back_end/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
cryptography==3.4.7
Django==3.2.5
django-cors==0.1
django-cors-headers==3.8.0
django-rest-knox==4.1.0
djangorestframework==3.12.4
djongo==1.3.6
dnspython==2.1.0
Pillow==8.3.1
pycparser==2.20
pymongo==3.12.0
pytz==2021.1
sqlparse==0.2.4
typing-extensions==3.10.0.0
